+++ .eslintignore
@@ -1,3 +1,4 @@
 node_modules/**
 .storybook/**
 public/**
+_story.js
