+++ src/components/FileUp/_story.js
@@ -1,0 +1,30 @@
+import React from 'react';
+import { storiesOf } from '@storybook/react';
+import { specs, describe, it } from 'storybook-addon-specifications';
+
+import { mount } from 'enzyme';
+import expect from 'expect';
+
+import { withKnobs, text, number } from '@storybook/addon-knobs';
+import { withInfo } from '@storybook/addon-info';
+import FileUp from './FileUp';
+
+storiesOf('File Upload', module)
+  .addDecorator(withKnobs)
+  .add('Single',
+    withInfo('File upload')(() => {
+      const maxFiles = number('Max Files', 5);
+      const maxSize = number('Max Size', 2000000);
+      const types = text('Types', 'image/*, application/pdf');
+      const story = <FileUp maxFiles={maxFiles} maxSize={maxSize} types={types} onChange={() => {}} />;
+
+      specs(() => describe('File Upload', () => {
+        it('Should have a label and "click to upload" in it', () => {
+          const output = mount(story);
+          expect(output.find('label').text()).toContain('click to upload');
+        });
+      }));
+
+      return story;
+    }),
+  );
