+++ src/components/GrantsNearYou/_story.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import { storiesOf } from '@storybook/react';
+
+import { withInfo } from '@storybook/addon-info';
+import GrantsNearYou from './GrantsNearYou';
+
+storiesOf('Grants Near You', module)
+  .add('GrantsNearYou',
+    withInfo('doc string about my component')(() => {
+      const SEARCH = 'https://1kfs7evxca.execute-api.eu-west-1.amazonaws.com/beta/grants-geo';
+      const POSTCODE_API = 'https://api.postcodes.io';
+      return (<GrantsNearYou searchURL={SEARCH} postcodeAPI={POSTCODE_API} />);
+    }),
+  );
