+++ src/components/SchoolsLookUp/package.json
@@ -1,4 +1,3 @@
-
 {
   "name": "SchoolsLookUp",
   "version": "0.0.0",
@@ -10,5 +9,8 @@
     "react": "15",
     "react-bootstrap-typeahead": "^2.0.0-alpha.4",
     "react-fontawesome": "1"
+  },
+  "devDependencies": {
+    "@storybook/react": "^3.4.8"
   }
 }
